+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 ### NEXT
 
 * Fix `SeqManager`: Properly consider previous cycle dropped inputs ([PR #1032](https://github.com/versatica/mediasoup/pull/1032)).
+* `RtpRetransmissionBuffer`: Get rid of not necessary `startSeq` private member ([PR #1029](https://github.com/versatica/mediasoup/pull/1029).
+* Node: Upgrade TypeScript to 5.0.2.
+
+
+### 3.11.15
+
+* `RtpRetransmissionBuffer`: Fix crash and add fuzzer ([PR #1028](https://github.com/versatica/mediasoup/pull/1028).
+* Update NPM deps.
 
 
 ### 3.11.14